backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
email-validator==2.2.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.1

